added some command line sugar to the recorder

+++ recorder/recorder.js
@@ -16,9 +16,35 @@ var WebSocket = require('ws');
 var ws;
 
 var samples = [];
+var sampleCount = 0;
 var fs = require('fs');
 
 var strftime = require('strftime')
+
+// options
+// arg[0] is node
+// arg[1] is the pwd
+// hence we are interested in index > 1
+
+if ((process.argv.indexOf('-h') > 1) || (process.argv.indexOf('--help') > 1)) {
+	console.log('space walk recorder help:');
+	console.log('The recorder records session data to a json file named automatically.');
+	console.log('usage: ./recorder.js [flags]');
+	console.log('\t -h \t this help message');
+	console.log('\t -v \t verbose output');
+	console.log('\t -d \t dry run, i.e. don\' write to the file system');
+	process.exit();
+}
+
+var verbose = (process.argv.indexOf('-v') > 1);
+if (verbose) {
+	console.log('verbose!');
+}
+
+var dryRun = (process.argv.indexOf('-d') > 1);
+if (dryRun) {
+	console.log('This is a dry run, no data is saved to disk!!!');
+}
 
 
 console.log('starting recorder...');
@@ -32,13 +58,18 @@ var sessionEnd = function() {
 
 	var fileName = strftime('%F_%H-%M-%S');
 
-	fs.writeFile('sessions/positions_' + fileName + '.json', JSON.stringify(samples, null, '\t'), function(err) {
-		if (err) {
-			console.log('error writing to file: ' + err);
-		} else {
-			console.log('+++ json file writte: ' + fileName);
-		}
-	})
+	if (!dryRun) {
+		fs.writeFile('sessions/positions_' + fileName + '.json', JSON.stringify(samples, null, '\t'), function(err) {
+			if (err) {
+				console.log('error writing to file: ' + err);
+			} else {
+				console.log('+++ json file writte: ' + fileName);
+			}
+		})
+	}
+
+	samples = [];
+	sampleCount = 0;
 }
 
 var keepAlive = function() {
@@ -83,8 +114,13 @@ function connect() {
 		
 		ws.onmessage = function(msg) {
 			keepAlive();
-			console.log('message got: ' + msg.data);
+			
+			if (verbose) {
+				console.log('message got: ' + msg.data);
+			}
+
 			samples.push(JSON.parse(msg.data));
+			sampleCount += 1;
 		}
 	}
 }
@@ -104,6 +140,17 @@ var checkSession = function() {
 }
 
 setInterval(checkSession, 1000);
+
+var showProgress = function() {
+	if (sessionActive) {
+		console.log('samples received: ' + sampleCount);
+	}
+	else {
+		console.log('waiting for session to begin');
+	}
+}
+
+setInterval(showProgress, 3000);
